new - Using Nigel's Solution
Nigel says:

The easy workaround is to always use new with a typename. So you declare a name for the array type you need and all should be OK.
If you want, you can change the grammar to add a new expression construct something like this:
   'new' '(' type ')'
It has the incorrect side-effect of making 'new' into a keyword, but we won't worry about that.

Going with the latter solution.